Issue #985 Move reducer to its own file
