+++ src/application/constants.ts
@@ -93,3 +93,13 @@ export const CLOSE_ABOUT_MODAL = 'CLOSE_ABOUT_MODAL';
 export const OPEN_ABOUT_MODAL = 'OPEN_ABOUT_MODAL';
 export const CLOSE_DISCLAIMER_MODAL = 'CLOSE_DISCLAIMER_MODAL';
 export const OPEN_DISCLAIMER_MODAL = 'OPEN_DISCLAIMER_MODAL';
+
+export const SUGGEST_AN_UPDATE = 'SUGGEST_AN_UPDATE';
+export const CHOOSE_CHANGE_NAME_AND_DETAILS = 'CHOOSE_CHANGE_NAME_AND_DETAILS';
+export const CHOOSE_REMOVE_SERVICE = 'CHOOSE_REMOVE_SERVICE';
+export const CHOOSE_OTHER_CHANGES = 'CHOOSE_OTHER_CHANGES';
+export const DISCARD_CHANGES = 'DISCARD_CHANGES';
+export const CLOSE = 'CLOSE';
+export const BACK = 'BACK';
+export const SUBMIT = 'SUBMIT';
+export const FINISH_FEEDBACK = 'FINISH_FEEDBACK';
